Indicate resource type with color bar in regular resource headers
