+++ app/src/main/res/layout/list_item_resource_header_card.xml
@@ -15,22 +15,31 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        android:background="@drawable/ripple_white_lightgray"
-        android:padding="16dp">
+        android:background="@drawable/ripple_white_lightgray" >
+
+        <include layout="@layout/resource_color_bar"
+            android:id="@+id/resourceColorBar"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"/>
 
         <TextView
             android:id="@+id/titleTextView"
             style="@style/HeaderCardTitleText"
             android:paddingBottom="12dp"
             tools:text="Lesson Title"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintTop_toTopOf="parent"/>
+            app:layout_constraintTop_toBottomOf="@id/resourceColorBar"/>
 
         <TextView
             android:id="@+id/authorNameTextView"
             style="@style/HeaderCardSubheaderText"
             tools:text="Author Name Here"
             android:paddingBottom="6dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/titleTextView" />
 
@@ -39,6 +48,8 @@
             style="@style/HeaderCardSubheaderText"
             tools:text="Institution Name Here"
             android:paddingBottom="6dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/authorNameTextView" />
 
@@ -47,9 +58,9 @@
             style="@style/HeaderCardSubheaderText"
             tools:text="Location Name Here"
             android:paddingBottom="6dp"
-            app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/institutionTextView"
-            app:layout_constraintRight_toLeftOf="@+id/submissionCountTextView"/>
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toStartOf="@+id/submissionCountTextView"/>
 
         <TextView
             android:id="@+id/dateEditedTextView"
@@ -57,8 +68,8 @@
             android:maxLines="1"
             tools:text="1 December, 2017"
             app:layout_constraintTop_toBottomOf="@id/locationTextView"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toLeftOf="@+id/submissionCountTextView"/>
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toStartOf="@+id/submissionCountTextView"/>
 
         <TextView
             android:id="@+id/submissionCountTextView"
@@ -74,10 +85,13 @@
             android:maxLines="1"
             android:ellipsize="end"
             android:padding="12dp"
+            android:layout_marginRight="16dp"
+            android:layout_marginEnd="16dp"
+            android:layout_marginBottom="16dp"
             android:background="@drawable/gray_ripple_pill"
             android:textColor="@color/gray"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
             tools:text="+50"/>
 
     </android.support.constraint.ConstraintLayout>
