+++ app/src/main/res/layout/resource_color_bar.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<View android:layout_width="0dp"
+    android:layout_height="6dp"
+    xmlns:android="http://schemas.android.com/apk/res/android" />
