+++ app/src/main/res/values/styles.xml
@@ -63,6 +63,8 @@
     <style name="HeaderCardText">
         <item name="android:layout_height">wrap_content</item>
         <item name="android:layout_width">0dp</item>
+        <item name="android:layout_marginLeft">16dp</item>
+        <item name="android:layout_marginRight">16dp</item>
         <item name="android:ellipsize">end</item>
         <item name="android:gravity">center_vertical</item>
         <item name="android:lineSpacingMultiplier">1.15</item>
@@ -73,6 +75,7 @@
         <item name="android:layout_height">60dp</item>
         <item name="fontPath">fonts/Roboto-Medium.ttf</item>
         <item name="android:textColor">@color/colorPrimary</item>
+        <item name="android:layout_marginTop">16dp</item>
         <item name="android:textStyle">bold</item>
         <item name="android:textSize">18sp</item>
         <item name="android:maxLines">2</item>
